% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{add_node_tip}
\alias{add_node_tip}
\title{Add one tip}
\usage{
add_node_tip(tree, host, time, label, state)
}
\arguments{
\item{tree}{transmission tree of class \code{treedata}, result of \code{\link{getTransmissionTree}}}

\item{host}{ID of the host}

\item{time}{time of the sampling event}

\item{label}{label of the new tip to be added}

\item{state}{state of the sampled individual}
}
\value{
modified tree, with on extra tip of length zero at the right place of the tree
}
\description{
Add a tip to the transmission tree corresponding to the sampled individual.
}
\keyword{internal}
