% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{applyFunctionToHosts}
\alias{applyFunctionToHosts}
\title{Apply a function to table.host}
\usage{
applyFunctionToHosts(res, pres.time, pasedFunction, active.hosts)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}

\item{pres.time}{current time}

\item{pasedFunction}{parsed exit/moving function}

\item{active.hosts}{a boolean vector of active hosts}
}
\value{
result vector
}
\description{
Return a vector of the results of the function
}
\keyword{internal}
