% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks-Messages.R
\name{CoreSanityChecks}
\alias{CoreSanityChecks}
\title{Checks if the simulator can start}
\usage{
CoreSanityChecks(length.sim, max.infected, init.individuals)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{number of initially infected individuals.}
}
\description{
Checks if the simulator can start: did the user provide a length, a maximum number of infected individuals and a number of "seeding" individuals?
}
\keyword{internal}
