% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{cumulativeInfected}
\alias{cumulativeInfected}
\title{Cumulative number of infected hosts at time t}
\usage{
cumulativeInfected(table.nosoi, t)
}
\arguments{
\item{table.nosoi}{result of function \code{\link{nosoiSim}}}

\item{t}{time (integer)}
}
\value{
Number of infected units at time t
}
\description{
For a given time t, this function returns the cumulative number of infected units since the start of the simulation.
}
\seealso{
\code{\link{nosoiSim}}
}
\keyword{internal}
