% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{drawBernouilli}
\alias{drawBernouilli}
\title{Draw newly infected}
\usage{
drawBernouilli(p)
}
\arguments{
\item{p}{vector of size K, giving the probability that each encounter
leads to an infection.}
}
\value{
Boolean vector giving the newly infected individuals.
}
\description{
For each encounter, simulate whether a new individual is infected.
}
\keyword{internal}
