% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-discrete.R
\name{dualDiscrete}
\alias{dualDiscrete}
\title{Dual-host with structured host population}
\usage{
dualDiscrete(length.sim, max.infected.A, max.infected.B,
  init.individuals.A, init.individuals.B, init.structure.A,
  init.structure.B, structure.matrix.A, structure.matrix.B, pExit.A,
  param.pExit.A, timeDep.pExit.A = FALSE, diff.pExit.A = FALSE,
  pMove.A, param.pMove.A, timeDep.pMove.A = FALSE,
  diff.pMove.A = FALSE, timeContact.A, param.timeContact.A,
  timeDep.timeContact.A = FALSE, diff.timeContact.A = FALSE, pTrans.A,
  param.pTrans.A, timeDep.pTrans.A = FALSE, diff.pTrans.A = FALSE,
  prefix.host.A = "H", pExit.B, param.pExit.B, timeDep.pExit.B = FALSE,
  diff.pExit.B = FALSE, pMove.B, param.pMove.B,
  timeDep.pMove.B = FALSE, diff.pMove.B = FALSE, timeContact.B,
  param.timeContact.B, timeDep.timeContact.B = FALSE,
  diff.timeContact.B = FALSE, pTrans.B, param.pTrans.B,
  timeDep.pTrans.B = FALSE, diff.pTrans.B = FALSE,
  prefix.host.B = "V", progress.bar = TRUE, print.step = 10)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected.A}{specifies the maximum number of hosts A that can be infected in the simulation.}

\item{max.infected.B}{specifies the maximum number of hosts B that can be infected in the simulation.}

\item{init.individuals.A}{number of initially infected individuals (hosts A).}

\item{init.individuals.B}{number of initially infected individuals (hosts B).}

\item{init.structure.A}{which state (e.g. location) the initially infected individuals of host A are located. (NA if init.individual.A is 0)}

\item{init.structure.B}{which state (e.g. location) the initially infected individuals of host B are located. (NA if init.individual.B is 0)}

\item{structure.matrix.A}{transition matrix (probabilities) to go from location A (row) to B (column) for host A.}

\item{structure.matrix.B}{transition matrix (probabilities) to go from location A (row) to B (column) for host B.}

\item{pExit.A}{function that gives the probability to exit the simulation for an infected host A (either moving out, dying, etc.).}

\item{param.pExit.A}{parameter names (list of functions) for the pExit for host A.}

\item{timeDep.pExit.A}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.pExit.A}{is pExit different between states of the structured population (TRUE/FALSE) for host A.}

\item{pMove.A}{function that gives the probability of a host moving as a function of time for host A.}

\item{param.pMove.A}{parameter names (list of functions) for the pMove for host A.}

\item{timeDep.pMove.A}{is moveDist dependant on the absolute time of the simulation (TRUE/FALSE) for host A.}

\item{diff.pMove.A}{is pMove different between states of the structured population (TRUE/FALSE) for host A.}

\item{timeContact.A}{function that gives the number of potential transmission events per unit of time  for host A.}

\item{param.timeContact.A}{parameter names (list of functions) for param.timeContact  for host A.}

\item{timeDep.timeContact.A}{is timeContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.timeContact.A}{is timeContact different between states of the structured population (TRUE/FALSE) for host A.}

\item{pTrans.A}{function that gives the probability of transmit a pathogen as a function of time since infection  for host A.}

\item{param.pTrans.A}{parameter names (list of functions) for the pExit  for host A.}

\item{timeDep.pTrans.A}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.pTrans.A}{is pTrans different between states of the structured population (TRUE/FALSE) for host A.}

\item{prefix.host.A}{character(s) to be used as a prefix for the host A identification number.}

\item{pExit.B}{function that gives the probability to exit the simulation for an infected host B (either moving out, dying, etc.).}

\item{param.pExit.B}{parameter names (list of functions) for the pExit for host B.}

\item{timeDep.pExit.B}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.pExit.B}{is pExit different between states of the structured population (TRUE/FALSE) for host B.}

\item{pMove.B}{function that gives the probability of a host moving as a function of time for host B.}

\item{param.pMove.B}{parameter names (list of functions) for the pMove for host B.}

\item{timeDep.pMove.B}{is moveDist dependant on the absolute time of the simulation (TRUE/FALSE) for host B.}

\item{diff.pMove.B}{is pMove different between states of the structured population (TRUE/FALSE) for host B.}

\item{timeContact.B}{function that gives the number of potential transmission events per unit of time  for host B.}

\item{param.timeContact.B}{parameter names (list of functions) for param.timeContact  for host B.}

\item{timeDep.timeContact.B}{is timeContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.timeContact.B}{is timeContact different between states of the structured population (TRUE/FALSE) for host B.}

\item{pTrans.B}{function that gives the probability of transmit a pathogen as a function of time since infection  for host B.}

\item{param.pTrans.B}{parameter names (list of functions) for the pExit  for host B.}

\item{timeDep.pTrans.B}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.pTrans.B}{is pTrans different between states of the structured population (TRUE/FALSE) for host B.}

\item{prefix.host.B}{character(s) to be used as a prefix for the host B identification number.}

\item{progress.bar}{if TRUE, displays a progress bar (current time/length.sim).}

\item{print.step}{progress.bar is TRUE, step with which the progress message will be printed.}
}
\description{
This function runs a single-host transmission chain simulation, without any spatial features. The simulation stops either at
the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.
}
