% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-none.R
\name{dualNone}
\alias{dualNone}
\title{Dual-host without structured host population}
\usage{
dualNone(length.sim, max.infected.A, max.infected.B, init.individuals.A,
  init.individuals.B, pExit.A, param.pExit.A, timeDep.pExit.A = FALSE,
  nContact.A, param.nContact.A, timeDep.nContact.A = FALSE, pTrans.A,
  param.pTrans.A, timeDep.pTrans.A = FALSE, prefix.host.A = "H",
  pExit.B, param.pExit.B, timeDep.pExit.B = FALSE, nContact.B,
  param.nContact.B, timeDep.nContact.B = FALSE, pTrans.B, param.pTrans.B,
  timeDep.pTrans.B = FALSE, prefix.host.B = "V",
  print.progress = TRUE, print.step = 10)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected.A}{specifies the maximum number of hosts A that can be infected in the simulation.}

\item{max.infected.B}{specifies the maximum number of hosts B that can be infected in the simulation.}

\item{init.individuals.A}{number of initially infected individuals (hosts A).}

\item{init.individuals.B}{number of initially infected individuals (hosts B).}

\item{pExit.A}{function that gives the probability to exit the simulation for an infected host A (either moving out, dying, etc.).}

\item{param.pExit.A}{parameter names (list of functions) for the pExit for host A.}

\item{timeDep.pExit.A}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{nContact.A}{function that gives the number of potential transmission events per unit of time  for host A.}

\item{param.nContact.A}{parameter names (list of functions) for param.nContact  for host A.}

\item{timeDep.nContact.A}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{pTrans.A}{function that gives the probability of transmit a pathogen as a function of time since infection  for host A.}

\item{param.pTrans.A}{parameter names (list of functions) for the pExit  for host A.}

\item{timeDep.pTrans.A}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{prefix.host.A}{character(s) to be used as a prefix for the host A identification number.}

\item{pExit.B}{function that gives the probability to exit the simulation for an infected host B (either moving out, dying, etc.).}

\item{param.pExit.B}{parameter names (list of functions) for the pExit for host B.}

\item{timeDep.pExit.B}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{nContact.B}{function that gives the number of potential transmission events per unit of time  for host B.}

\item{param.nContact.B}{parameter names (list of functions) for param.nContact  for host B.}

\item{timeDep.nContact.B}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{pTrans.B}{function that gives the probability of transmit a pathogen as a function of time since infection  for host B.}

\item{param.pTrans.B}{parameter names (list of functions) for the pExit  for host B.}

\item{timeDep.pTrans.B}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{prefix.host.B}{character(s) to be used as a prefix for the host B identification number.}

\item{print.progress}{if TRUE, displays a progress bar (current time/length.sim).}

\item{print.step}{print.progress is TRUE, step with which the progress message will be printed.}
}
\description{
This function runs a dual-host transmission chain simulation, without any spatial features. The simulation stops either at
the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.
}
