% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{endMessage}
\alias{endMessage}
\title{End message}
\usage{
endMessage(Host.count, pres.time)
}
\arguments{
\item{Host.count}{number of infected hosts}

\item{pres.time}{current time of the simulation}
}
\description{
Message that ends the simulation
}
\keyword{internal}
