% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks-Messages.R
\name{FunctionSanityChecks}
\alias{FunctionSanityChecks}
\title{Checks if a function is properly formatted}
\usage{
FunctionSanityChecks(pFunc, name, param.pFunc, timeDep, diff, hostCount,
  continuous, stateNames)
}
\arguments{
\item{pFunc}{a function}

\item{name}{the name of the function}

\item{param.pFunc}{a named list of arguments}

\item{timeDep}{is the function differential according to absolute time? (TRUE/FALSE)}

\item{diff}{is the function differential according to state/env.variable? (TRUE/FALSE)}

\item{hostCount}{is the function differential according to host count? (TRUE/FALSE)}

\item{continuous}{is the function to be used in a continuous space? (TRUE/FALSE)}

\item{stateNames}{name of the states (vector) in case of discrete structure.}

\item{structure}{is the function to be used in a structured population? (TRUE/FALSE)}
}
\description{
Checks if the function is properly formatted given the user's input and nosoi's requirements.
}
\keyword{internal}
