% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{get_node}
\alias{get_node}
\title{Get Node}
\usage{
get_node(tdata, host, time)
}
\arguments{
\item{tdata}{a tibble extracted from a \code{treedata} object.}

\item{host}{host id}

\item{time}{time at which to sample the host}
}
\value{
number of the node in the transmission tree descending from the
host at given time.
}
\description{
Find the node descending from the branch where the host lies at a given
 time on the transmission tree.
}
\seealso{
\code{\link{get_position}}
}
\keyword{internal}
