% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{get_other_count}
\alias{get_other_count}
\title{Get host count from table}
\usage{
get_other_count(table_hosts, name_cur, active_hosts, name_count, cur)
}
\arguments{
\item{table_hosts}{a table.hosts}

\item{name_cur}{name of the cell, one of "current.in" or "current.cell.raster"}

\item{active_hosts}{a boolean vector of active hosts}

\item{name_count}{name of the count, one of "host.count.A" or "host.count.B"}

\item{cur}{a data.table with the state}
}
\value{
data.table containing state.ID (col1) and count (col2)
}
\description{
Returns the host count for a table and a state
}
\keyword{internal}
