% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{get_position}
\alias{get_position}
\title{Get Position on branch}
\usage{
get_position(tdata, node, time)
}
\arguments{
\item{tdata}{a tibble extracted from a \code{treedata} object.}

\item{node}{node below the sampling event}

\item{time}{time of the sampling event}
}
\value{
time between the node and the sampling event.
}
\description{
Find the position on the branch above the node of the times sample.
 Warning: The node needs to be extracted with \code{\link{get_node}}.
 Result of this function is to be used in \code{\link[ape]{bind.tree}}
}
\seealso{
\code{\link{get_node}}
}
\keyword{internal}
