% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{getCumulative}
\alias{getCumulative}
\title{Cumulative number of infected units for the full length of the simulation}
\usage{
getCumulative(table.nosoi)
}
\arguments{
\item{table.nosoi}{result of function \code{\link{nosoiSim}}}
}
\description{
Number of infected units at time t for the full length of the simulation
}
\seealso{
\code{\link{nosoiSim}}
}
