% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExiting.R
\name{getExiting}
\alias{getExiting}
\title{Get Exiting individuals}
\usage{
getExiting(res, pres.time, pExitParsed)
}
\arguments{
\item{res}{an object of class \code{nosoiSim}.}

\item{pres.time}{current time}

\item{pExitParsed}{parsed exit function}
}
\value{
Boolean vector
}
\description{
Return a vector of exiting individuals.
}
\keyword{internal}
