% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{getExitingMoving}
\alias{getExitingMoving}
\title{Get Exiting or Moving individuals}
\usage{
getExitingMoving(res, pres.time, pasedFunction)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}

\item{pres.time}{current time}

\item{pasedFunction}{parsed exit/moving function}
}
\value{
Boolean vector
}
\description{
Return a vector of exiting individuals.
}
\keyword{internal}
