% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utilityFunctions.R
\name{getHostInfo}
\alias{getHostInfo}
\title{get host info}
\usage{
getHostInfo(res, what = c("table.hosts", "N.infected", "table.state",
  "popStructure"), pop = "A")
}
\arguments{
\item{res}{an object of class \code{nosoiSim}}

\item{what}{the information to get. One of "table.hosts", "N.infected", "table.state", "popStructure"}

\item{pop}{the population to be extracted (one of "A" or "B")}
}
\value{
a data.table with host informations
}
\description{
Creates a \code{nosoiSim} object.
}
