% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{getPositionInfected}
\alias{getPositionInfected}
\title{get Position Infected}
\usage{
getPositionInfected(nosoiSim, df.meetTransmit, i)
}
\arguments{
\item{nosoiSim}{an object of class \code{nosoiSim}.}

\item{df.meetTransmit}{current data.table for transmission}

\item{i}{individual}
}
\value{
A vector of positions
}
\description{
Return the relevant position of the infected individual.
}
\keyword{internal}
