% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{getR0}
\alias{getR0}
\title{Get R0 from a nosoi simulation}
\usage{
getR0(table.nosoi)
}
\arguments{
\item{table.nosoi}{result of function \code{\link{nosoiSim}}}
}
\value{
R0 estimate (mean and distribution), as well as the number of inactive hosts at the end of the simulation.
}
\description{
Gets an estimate of secondary cases (what R0 usually tries to estimate) and its distribution. The calculation is based on inactive hosts at the end of the simulation.
}
\seealso{
\code{\link{nosoiSim}}
}
