% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utilityFunctions.R
\name{getTableHosts}
\alias{getTableHosts}
\title{Get table hosts}
\usage{
getTableHosts(res, pop = "A")
}
\arguments{
\item{res}{an object of class \code{nosoiSim}}

\item{pop}{the population to be extracted (one of "A" or "B")}
}
\value{
a data.table with hosts table informations
}
\description{
Get "table.hosts"
TODO: describe the table
}
