% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{getTableState}
\alias{getTableState}
\title{Get table states}
\usage{
getTableState(res, pop = "A")
}
\arguments{
\item{res}{an object of class \code{nosoiSim}}

\item{pop}{the population to be extracted (one of "A" or "B")}
}
\value{
a data.table with state table informations
}
\description{
Get "table.state"
TODO: describe the table
}
