% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniTable.R
\name{iniTable}
\alias{iniTable}
\title{Generates initial table to start the simulation (internal fonction)}
\usage{
iniTable(init.individuals, init.structure, prefix.host, param.pExit,
  param.pMove, param.timeContact, param.pTrans,
  current.environmental.value = NA)
}
\arguments{
\item{init.individuals}{number of initially infected individuals (i.e. number of lines at time 0).}

\item{init.structure}{State (or coordinates) of the initially infected individuals.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{param.pExit}{list of exit probability function(s).}

\item{param.pMove}{list of movement probability function(s).}

\item{param.timeContact}{list of contact probability function(s).}

\item{param.pTrans}{list of transmission probability function(s).}

\item{current.environmental.value}{current value of the environemental variable provided by the raster according to its position in init.structure.}
}
\description{
This function creates the initial table for the host.
}
\keyword{internal}
