% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniTableState.R
\name{iniTableState}
\alias{iniTableState}
\title{Generates initial movement table to start the simulation (internal fonction)}
\usage{
iniTableState(init.individuals, init.structure, prefix.host,
  current.environmental.value = NA)
}
\arguments{
\item{init.individuals}{number of initially infected individuals (i.e. number of lines at time 0).}

\item{init.structure}{State of the initially infected individuals.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{current.environmental.value}{current value of the environemental variable provided by the raster according to its position in init.structure.}
}
\description{
This function creates the initial table for the host, with 5+number of parameters of the transmission probability function paramters, and init.individuals row(s).
}
\keyword{internal}
