% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{keep.tip.treedata}
\alias{keep.tip.treedata}
\title{Keep tips}
\usage{
keep.tip.treedata(tree, tip)
}
\description{
Keep the tips in the list. See \code{\link[ape:keep.tip]{keep.tip}}
}
\keyword{internal}
