% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{keepState}
\alias{keepState}
\title{Should we build the table.host table}
\usage{
keepState(nosoiSim)
}
\arguments{
\item{nosoiSim}{an object of class \code{nosoiSim}.}
}
\value{
boolean
}
\description{
Should we build the table.host table
}
\keyword{internal}
