% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoves.R
\name{makeMoves}
\alias{makeMoves}
\title{Make Move function}
\usage{
makeMoves(res, pres.time, moving.full, structure.matrix = NA,
  moveDistParsed = NA, structure.raster = NA,
  attracted.by.raster = NA, max.raster = NA)
}
\arguments{
\item{res}{an object of class \code{nosoiSim}.}

\item{pres.time}{current time}

\item{structure.matrix}{transition matrix (probabilities) to go from location A (row) to B (column)}

\item{structure.raster}{raster object defining the environmental variable.}

\item{attracted.by.raster}{should the hosts be attracted by high values in the environmental raster? (TRUE/FALSE)}
}
\value{
The modified object res
}
\description{
Makes the move.
}
\keyword{internal}
