% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks.R
\name{MatrixSanityChecks}
\alias{MatrixSanityChecks}
\title{Checks if the matrix is properly formated}
\usage{
MatrixSanityChecks(structure.matrix, init.structure)
}
\arguments{
\item{structure.matrix}{transition matrix (probabilities) to go from location A (row) to B (column)}

\item{init.structure}{which state (i.e. location) the initially infected individuals are located.}
}
\description{
Checks if the transition matrix is properly formated given the user's input and nosoi's requirements.
}
\keyword{internal}
