% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetTransmit.R
\name{meetTransmit}
\alias{meetTransmit}
\title{Meet Transmit function}
\usage{
meetTransmit(res, pres.time, active.hosts, positions, timeContactParsed,
  pTransParsed, prefix.host, param.pExit, param.pMove, param.timeContact,
  param.pTrans, param.moveDist)
}
\arguments{
\item{res}{an object of class \code{nosoiSim}.}

\item{param.pExit}{list of exit probability function(s).}

\item{param.pMove}{list of movement probability function(s).}

\item{param.pTrans}{list of transmission probability function(s).}
}
\value{
The modified object res
}
\description{
Perform the tasks related to meeting and transmission events.
}
\keyword{internal}
