% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{merge_host_tables}
\alias{merge_host_tables}
\alias{merge_state_tables}
\title{Merge Population Data}
\usage{
merge_host_tables(nosoiInf)

merge_state_tables(nosoiInf)
}
\arguments{
\item{nosoiInf}{an object of class \code{\link{nosoiSim}}}
}
\value{
the merged `table.hosts` or `table.state` of the two populations.
}
\description{
Utility function to get merged hosts data if dual.
}
\keyword{internal}
