% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_step1-makeMoves.R
\name{moveFunction.continuous}
\alias{moveFunction.continuous}
\title{Make Continuous Move function}
\usage{
moveFunction.continuous(res, pres.time, moving.full, sdMoveParsed,
  structure.raster, attracted.by.raster, max.raster, ...)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}

\item{pres.time}{current time}

\item{structure.raster}{raster object defining the environmental variable.}

\item{attracted.by.raster}{should the hosts be attracted by higher values in the environmental raster? (TRUE/FALSE).}
}
\value{
The modified object res
}
\description{
Makes the continuous move.
}
\keyword{internal}
