% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoves.R
\name{moveFunction.singleContinuous}
\alias{moveFunction.singleContinuous}
\title{Make Continuous Move function}
\usage{
moveFunction.singleContinuous(res, pres.time, moving.full, moveDistParsed,
  structure.raster, attracted.by.raster, max.raster, ...)
}
\arguments{
\item{res}{an object of class \code{nosoiSim}.}

\item{pres.time}{current time}

\item{structure.raster}{raster object defining the environmental variable.}

\item{attracted.by.raster}{should the hosts be attracted by high values in the environmental raster? (TRUE/FALSE)}

\item{...}{other arguments to be passed on to the simulator (see below).}
}
\value{
The modified object res
}
\description{
Makes the continuous move.
}
\keyword{internal}
