% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_step1-makeMoves.R
\name{moveRotateContinuous}
\alias{moveRotateContinuous}
\title{Rotates movement in 2D space}
\usage{
moveRotateContinuous(pt1, dX, dY, angle)
}
\arguments{
\item{pt1}{current position (x,y)}

\item{dX}{movement in X dim}

\item{dY}{movement in Y space}

\item{angle}{angle}
}
\description{
Rotates movement in 2D space
}
\keyword{internal}
