% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_tablesManagment.R
\name{newLine}
\alias{newLine}
\title{Creates a new line to be added to the table when new host is infected (internal fonction)}
\usage{
newLine(hosts.ID, infected.by, infected.in, time.is, ParamHost,
  current.environmental.value = NULL)
}
\arguments{
\item{hosts.ID}{unique ID for the new host}

\item{infected.by}{unique ID of host that transmits to the new one}

\item{infected.in}{state in which the host was infected}

\item{time.is}{time in the simulation, when the infection takes place}

\item{ParamHost}{list of individual based parameters.}

\item{current.in}{state in which the host currently is}
}
\value{
a list with the new line to add.
}
\description{
This function creates a new line for the table.
The lines are to be bounded with \code{\link[data.table]{rbindlist}}.
}
\keyword{internal}
