% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newLineState.R
\name{newLineState}
\alias{newLineState}
\title{Creates a new line to be added to the movement table when hosts moves (internal fonction)}
\usage{
newLineState(hosts.ID, state.pres, time.is)
}
\arguments{
\item{hosts.ID}{unique ID for the new host}

\item{state.pres}{state in which host currently is}

\item{time.is}{time in the simulation, when the infection takes place}
}
\value{
a list with the new line to add.
}
\description{
This function creates a new line for the table,
The lines are to be bounded with \code{\link[data.table]{rbindlist}}.
}
\keyword{internal}
