% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{nosoiSimConstructor}
\alias{nosoiSimConstructor}
\title{nosoiSim Constructor}
\usage{
nosoiSimConstructor(N.infected, pres.time, table.hosts, table.state,
  type = c("singleNone", "singleDiscrete", "singleContinuous"))
}
\arguments{
\item{N.infected}{number of infected hosts}

\item{pres.time}{current time of the simulation}

\item{table.hosts}{data.table of hosts}

\item{table.state}{data.table of hosts movement}

\item{type}{name of the diffusion}
}
\description{
Creates a \code{nosoiSim} object.
}
\keyword{internal}
