% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{nosoiSimOneConstructor}
\alias{nosoiSimOneConstructor}
\title{nosoiSimOne Constructor}
\usage{
nosoiSimOneConstructor(N.infected, table.hosts, table.state, prefix.host,
  geoStructure = c("none", "discrete", "continuous"))
}
\arguments{
\item{N.infected}{number of infected hosts}

\item{table.hosts}{data.table of hosts}

\item{table.state}{data.table of hosts movement}

\item{geoStructure}{geographical structure (one of "none, "discrete" or "continuous")}
}
\value{
An object of class \code{nosoiSimOne}
}
\description{
Creates a \code{nosoiSim} object.
}
\keyword{internal}
