% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoiSummary.R
\name{nosoiSummary}
\alias{nosoiSummary}
\title{Summary of a simulation}
\usage{
nosoiSummary(nosoi.output)
}
\arguments{
\item{nosoi.output}{Output of a nosoi simulation (the data.table).}
}
\description{
This function provides summary informations about the simulation (number of infected hosts, R0, etc.).
}
\details{
All elements are provided in a list.
}
