% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{numberInfected}
\alias{numberInfected}
\title{Number of active infected hosts at time t}
\usage{
numberInfected(table.nosoi, t)
}
\arguments{
\item{table.nosoi}{result of function \code{\link{nosoiSim}}}

\item{t}{time (integer)}
}
\value{
Number of infected units at time t
}
\description{
For a given time t, this function returns the number of infected active units.
}
\seealso{
\code{\link{nosoiSim}}
}
\keyword{internal}
