% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{numberInfectedBGW}
\alias{numberInfectedBGW}
\title{Number of infected hosts at time t (BGW)}
\usage{
numberInfectedBGW(table.nosoi, t)
}
\arguments{
\item{table.nosoi}{result of function \code{\link{nosoiSim}}}

\item{t}{time (integer)}
}
\value{
Number of infected units at time t
}
\description{
For a given time t, this function returns the number of infected active units.
The difference with \code{\link{numberInfected}} is that it counts an "out"
individual as still there, but with no children.
This is for comparison with the BGW process, should be internal only.
}
\seealso{
\code{\link{nosoiSim}}
}
\keyword{internal}
