% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{numberInfectedStates}
\alias{numberInfectedStates}
\title{Number of active infected hosts at time t}
\usage{
numberInfectedStates(table.states, t)
}
\arguments{
\item{table.states}{state table, result of function \code{\link{nosoiSim}}}

\item{t}{time (integer)}
}
\value{
Number of infected units at time t
}
\description{
For a given time t, this function returns the number of infected active units.
}
\seealso{
\code{\link{nosoiSim}}
}
\keyword{internal}
