% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{outputWrapper}
\alias{outputWrapper}
\title{Output wrapper}
\usage{
outputWrapper(Host.count, pres.time, table.hosts, state.archive)
}
\arguments{
\item{state.archive}{}
}
\description{
Wraps up the nosoi output
}
\keyword{internal}
