% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFunction.R
\name{parseFunction}
\alias{parseFunction}
\title{Parse function for later use}
\usage{
parseFunction(pFunc, param.pFunc, name, diff = FALSE, timeDep = FALSE)
}
\arguments{
\item{pFunc}{a function}

\item{param.pFunc}{a named list of arguments}

\item{name}{the name of the function}

\item{diff}{is the function differential according to state/env.variable? (TRUE/FALSE)}
}
\value{
list of parsed quantities:
\itemize{
 \item{"type"}{A string, "simple" of "complex": are there several arguments to the function ?}
 \item{"nArgs"}{Number of arguments (besides time).}
 \item{"vect"}{Vectorized version of the function.}
 \item{"vectArgs"}{Vector of arguments for the vectorized function.}
}
}
\description{
Parse a user-provided function to get the vectorized version.
}
\keyword{internal}
