% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{progressMessage}
\alias{progressMessage}
\title{Progress bar}
\usage{
progressMessage(Host.count, pres.time, print.step, length.sim,
  max.infected)
}
\arguments{
\item{max.infected}{}
}
\description{
Echos the state of the simulation at any given time step provided by the user.
}
\keyword{internal}
