% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeGenerator.r
\name{sampleTransmissionTreeFromTheDead}
\alias{sampleTransmissionTreeFromTheDead}
\title{Sample the transmission tree among the dead}
\usage{
sampleTransmissionTreeFromTheDead(tree, hosts)
}
\arguments{
\item{tree}{a \code{\link[tidytree:treedata-class]{treedata}} object created by function \code{\link{getTransmissionTree}}}

\item{hosts}{a vector of dead hosts to sample}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree and the mapped data at all the nodes.
}
\description{
Sample a full transmission tree. This function allows for sampling only dead
 individuals (the sampling procedure is destructive).
}
\details{
The tree needs to be produced by function \code{\link{getTransmissionTree}}
 applied on the same \code{nosioSim} object.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

For sampling non-dead individuals, see \code{\link{sampleTransmissionTree}}.
}
