% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-none-nonparallel.R
\name{single_none_parallel}
\alias{single_none_parallel}
\title{Single-host, no spatial features and on a single core}
\usage{
single_none_parallel(length, max.infected, init.individuals, time_contact,
  p_trans, p_exit, ...)
}
\arguments{
\item{length}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{if simulation should be run in parallel or not.}

\item{time_contact}{function that gives the number of potential transmission events per unit of time.}

\item{p_trans}{function that gives the probability of transmit a pathogen as a function of time since infection.}

\item{p_exit}{probability to exit the simulation for an infected host (either moving out, dying, etc.).}

\item{...}{other arguments to be passed on to the simulator (see below).}
}
\description{
Single-host, no spatial features and on a single core
}
\details{
This function runs a single-host, without any spatial features, epidemiological simulation on a single core.

Th simulation stops either at the end of given time (specified by length) or when the number of hosts infected threshold is passed.
}
