% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-continuous.R
\name{singleContinuous}
\alias{singleContinuous}
\title{Single-host with structured (continuous) host population}
\usage{
singleContinuous(length.sim, max.infected, init.individuals,
  init.structure, structure.raster, diff.pMove = FALSE,
  timeDep.pMove = FALSE, pMove, param.pMove, diff.sdMove = FALSE,
  timeDep.sdMove = FALSE, sdMove, param.sdMove,
  attracted.by.raster = FALSE, diff.nContact = FALSE,
  timeDep.nContact = FALSE, nContact, param.nContact,
  diff.pTrans = FALSE, timeDep.pTrans = FALSE, pTrans, param.pTrans,
  diff.pExit = FALSE, timeDep.pExit = FALSE, pExit, param.pExit,
  prefix.host = "H", print.progress = TRUE, print.step = 10)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{number of initially infected individuals.}

\item{init.structure}{which state (i.e. location) the initially infected individuals are located.}

\item{structure.raster}{raster object defining the environmental variable.}

\item{diff.pMove}{is pMove different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pMove}{is pMove dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pMove}{function that gives the probability of a host moving as a function of time.}

\item{param.pMove}{parameter names (list of functions) for the pMove.}

\item{diff.sdMove}{is sdMove dependant on the environmental value (TRUE/FALSE).}

\item{timeDep.sdMove}{is sdMove dependant on the absolute time of the simulation (TRUE/FALSE).}

\item{sdMove}{function that gives the distance travelled (based on coordinates); gives the sd value for the brownian motion.}

\item{param.sdMove}{parameter names (list of functions) for sdMove}

\item{attracted.by.raster}{should the hosts be attracted by high values in the environmental raster? (TRUE/FALSE)}

\item{diff.nContact}{is nContact different between states of the structured population (TRUE/FALSE)}

\item{timeDep.nContact}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{nContact}{function that gives the number of potential transmission events per unit of time.}

\item{param.nContact}{parameter names (list of functions) for nContact.}

\item{diff.pTrans}{is pTrans different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pTrans}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pTrans}{function that gives the probability of transmit a pathogen as a function of time since infection.}

\item{param.pTrans}{parameter names (list of functions) for the pTrans.}

\item{diff.pExit}{is pExit different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pExit}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pExit}{function that gives the probability to exit the simulation for an infected host (either moving out, dying, etc.).}

\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{print.progress}{if TRUE, print message on simulation state to screen.}

\item{print.step}{print.progress is TRUE, step with which the progress message will be printed.}
}
\description{
This function runs a single-host transmission chain simulation, with a structured host population (such as spatial features).
The simulation stops either at the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected)
is passed.
}
