% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{updateHostCount}
\alias{updateHostCount}
\title{Summarise position of hosts in a discrete or discretized (raster) space}
\usage{
updateHostCount(res, res.B = NULL, type)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}
}
\value{
data.table containing state.ID (col1) and count (col2)
}
\description{
Returns a table of the number of hosts per discrete or discretized (raster cells) states.
}
\keyword{internal}
