% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_tablesManagment.R
\name{updateTableState}
\alias{updateTableState}
\title{Update table state with exiting individuals}
\usage{
updateTableState(res, exiting, pres.time)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}

\item{exiting}{a vector (TRUE/FALSE) of exiting individuals}

\item{pres.time}{current time}
}
\value{
\code{nosoiSim} object
}
\description{
Return a vector of exiting individuals.
}
\keyword{internal}
