% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_step2-MeetTransmit.R
\name{writeInfected}
\alias{writeInfected}
\title{Write newly infected function}
\usage{
writeInfected(df.meetTransmit, res, pres.time, ParamHost)
}
\arguments{
\item{df.meetTransmit}{table with active transmission events, coming from \code{meetTransmit}}

\item{res}{an object of class \code{nosoiSimOne}.}

\item{ParamHost}{list of individual based parameters.}
}
\value{
The modified object res
}
\description{
Perform the tasks related to creating new lines of hosts in relevant tables.
}
\keyword{internal}
