% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-discrete.R
\name{dualDiscrete}
\alias{dualDiscrete}
\title{Dual-host with structured (discrete) host population}
\usage{
dualDiscrete(length.sim, max.infected.A, max.infected.B,
  init.individuals.A, init.individuals.B, init.structure.A,
  init.structure.B, structure.matrix.A, structure.matrix.B, pExit.A,
  param.pExit.A, timeDep.pExit.A = FALSE, diff.pExit.A = FALSE,
  hostCount.pExit.A = FALSE, pMove.A, param.pMove.A,
  timeDep.pMove.A = FALSE, diff.pMove.A = FALSE,
  hostCount.pMove.A = FALSE, nContact.A, param.nContact.A,
  timeDep.nContact.A = FALSE, diff.nContact.A = FALSE,
  hostCount.nContact.A = FALSE, pTrans.A, param.pTrans.A,
  timeDep.pTrans.A = FALSE, diff.pTrans.A = FALSE,
  hostCount.pTrans.A = FALSE, prefix.host.A = "H", pExit.B,
  param.pExit.B, timeDep.pExit.B = FALSE, diff.pExit.B = FALSE,
  hostCount.pExit.B = FALSE, pMove.B, param.pMove.B,
  timeDep.pMove.B = FALSE, diff.pMove.B = FALSE,
  hostCount.pMove.B = FALSE, nContact.B, param.nContact.B,
  timeDep.nContact.B = FALSE, diff.nContact.B = FALSE,
  hostCount.nContact.B = FALSE, pTrans.B, param.pTrans.B,
  timeDep.pTrans.B = FALSE, diff.pTrans.B = FALSE,
  hostCount.pTrans.B = FALSE, prefix.host.B = "V",
  print.progress = TRUE, print.step = 10)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected.A}{specifies the maximum number of hosts A that can be infected in the simulation.}

\item{max.infected.B}{specifies the maximum number of hosts B that can be infected in the simulation.}

\item{init.individuals.A}{number of initially infected individuals (hosts A).}

\item{init.individuals.B}{number of initially infected individuals (hosts B).}

\item{init.structure.A}{which state (e.g. location) the initially infected individuals of host A are located. (NA if init.individual.A is 0)}

\item{init.structure.B}{which state (e.g. location) the initially infected individuals of host B are located. (NA if init.individual.B is 0)}

\item{structure.matrix.A}{transition matrix (probabilities) to go from location A (row) to B (column) for host A.}

\item{structure.matrix.B}{transition matrix (probabilities) to go from location A (row) to B (column) for host B.}

\item{pExit.A}{function that gives the probability to exit the simulation for an infected host A (either moving out, dying, etc.).}

\item{param.pExit.A}{parameter names (list of functions) for the pExit for host A.}

\item{timeDep.pExit.A}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.pExit.A}{is pExit different between states of the structured population (TRUE/FALSE) for host A.}

\item{hostCount.pExit.A}{does pExit varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pExit.A should be TRUE.}

\item{pMove.A}{function that gives the probability of a host moving as a function of time for host A.}

\item{param.pMove.A}{parameter names (list of functions) for the pMove for host A.}

\item{timeDep.pMove.A}{is pMove dependant on the absolute time of the simulation (TRUE/FALSE) for host A.}

\item{diff.pMove.A}{is pMove different between states of the structured population (TRUE/FALSE) for host A.}

\item{hostCount.pMove.A}{does pMove varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pMove.A should be TRUE.}

\item{nContact.A}{function that gives the number of potential transmission events per unit of time  for host A.}

\item{param.nContact.A}{parameter names (list of functions) for param.nContact  for host A.}

\item{timeDep.nContact.A}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.nContact.A}{is nContact different between states of the structured population (TRUE/FALSE) for host A.}

\item{hostCount.nContact.A}{does nContact varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.nContact.A should be TRUE.}

\item{pTrans.A}{function that gives the probability of transmit a pathogen as a function of time since infection  for host A.}

\item{param.pTrans.A}{parameter names (list of functions) for the pExit  for host A.}

\item{timeDep.pTrans.A}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host A.}

\item{diff.pTrans.A}{is pTrans different between states of the structured population (TRUE/FALSE) for host A.}

\item{hostCount.pTrans.A}{does pTrans varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pTrans.A should be TRUE.}

\item{prefix.host.A}{character(s) to be used as a prefix for the host A identification number.}

\item{pExit.B}{function that gives the probability to exit the simulation for an infected host B (either moving out, dying, etc.).}

\item{param.pExit.B}{parameter names (list of functions) for the pExit for host B.}

\item{timeDep.pExit.B}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.pExit.B}{is pExit different between states of the structured population (TRUE/FALSE) for host B.}

\item{hostCount.pExit.B}{does pExit varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pExit.B should be TRUE.}

\item{pMove.B}{function that gives the probability of a host moving as a function of time for host B.}

\item{param.pMove.B}{parameter names (list of functions) for the pMove for host B.}

\item{timeDep.pMove.B}{is pMove dependant on the absolute time of the simulation (TRUE/FALSE) for host B.}

\item{diff.pMove.B}{is pMove different between states of the structured population (TRUE/FALSE) for host B.}

\item{hostCount.pMove.B}{does pMove varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pMove.B should be TRUE.}

\item{nContact.B}{function that gives the number of potential transmission events per unit of time  for host B.}

\item{param.nContact.B}{parameter names (list of functions) for param.nContact  for host B.}

\item{timeDep.nContact.B}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.nContact.B}{is nContact different between states of the structured population (TRUE/FALSE) for host B.}

\item{hostCount.nContact.B}{does nContact varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.nContact.B should be TRUE.}

\item{pTrans.B}{function that gives the probability of transmit a pathogen as a function of time since infection  for host B.}

\item{param.pTrans.B}{parameter names (list of functions) for the pExit  for host B.}

\item{timeDep.pTrans.B}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)  for host B.}

\item{diff.pTrans.B}{is pTrans different between states of the structured population (TRUE/FALSE) for host B.}

\item{hostCount.pTrans.B}{does pTrans varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pTrans.B should be TRUE.}

\item{prefix.host.B}{character(s) to be used as a prefix for the host B identification number.}

\item{print.progress}{if TRUE, displays a progress bar (current time/length.sim).}

\item{print.step}{print.progress is TRUE, step with which the progress message will be printed.}
}
\description{
This function runs a dual-host transmission chain simulation, with discrete structure features. The simulation stops either at
the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.
}
