% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks-Messages.R
\name{endMessage}
\alias{endMessage}
\title{End message}
\usage{
endMessage(Host.count.A, Host.count.B = NULL, pres.time,
  type = "single")
}
\arguments{
\item{Host.count.A}{number of infected hosts (host A)}

\item{Host.count.B}{number of infected hosts (host B)}

\item{pres.time}{current time of the simulation}

\item{type}{either single/dual host}
}
\description{
Message that ends the simulation
}
\keyword{internal}
