% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nosoiSummary.R
\name{getCumulative}
\alias{getCumulative}
\title{Gets the cumulative number of infected hosts for the full length of the simulation}
\usage{
getCumulative(nosoi.output)
}
\arguments{
\item{nosoi.output}{Output of a nosoi simulation (object of class \code{\link{nosoiSim}}).}
}
\value{
The output is a \code{\link[data.table]{data.table}} with the following structure:
\describe{
   \item{t}{Time-step (integer).}
   \item{Count}{Cumulative number of infected hosts at given time-step.}
   \item{type}{Host-type, identified by its user-defined prefix.}
   }
}
\description{
This function computes from the output of a \code{nosoiSim} simulation the cumulative count of infected hosts at each time step of the simulation. The output is a \code{\link[data.table]{data.table}}.
}
\seealso{
\code{\link{summary.nosoiSim}}
}
