% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{getTransmissionTree}
\alias{getTransmissionTree}
\title{Get full transmission tree}
\usage{
getTransmissionTree(nosoiInf, pop = "A")
}
\arguments{
\item{nosoiInf}{an object of class \code{\link{nosoiSim}}}

\item{pop}{the population to be considered (one of "A" or "B")}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree and the mapped data at all the nodes.
}
\description{
From a simulated epidemics, this function extract the full transmission tree.
}
\details{
This function uses packages \pkg{tidytree} and \pkg{treeio},
 that rely on \code{\link[ape:ape-package]{ape}}.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

To sub-sample this tree, see function \code{\link{sampleTransmissionTree}}
}
