% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{getTransmissionTree}
\alias{getTransmissionTree}
\title{Gets the full transmission tree (phylogenetic tree-like) from a \code{nosoi} simulation}
\usage{
getTransmissionTree(nosoiInf)
}
\arguments{
\item{nosoiInf}{an object of class \code{\link{nosoiSim}}}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree based on the transmission chain and the mapped data at all the nodes.
}
\description{
From a \code{nosoi} simulated epidemics, this function extract the full transmission tree in a form mimicking a phylogenetic tree.
}
\details{
This function uses packages \pkg{tidytree} and \pkg{treeio},
 that rely on \code{\link[ape:ape-package]{ape}}.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

To sub-sample this tree, see function \code{\link{sampleTransmissionTree}}
}
