% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniTable.R
\name{iniTable}
\alias{iniTable}
\title{Generates initial table to start the simulation (internal fonction)}
\usage{
iniTable(init.individuals, prefix.host, n.pTrans.param, param.pTrans)
}
\arguments{
\item{init.individuals}{number of initially infected individuals (i.e. number of lines at time 0).}

\item{n.pTrans.param}{number of parameters for transmission probability computation (impacts number of columns).}

\item{pTrans.param}{transmission probability function parameters' name.}
}
\description{
Generates initial table to start the simulation (internal fonction)
}
\details{
This function creates the initial table for the host, with 5+number of parameters of the transmission probability function paramters, and init.individuals row(s).
}
