% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_tablesManagment.R
\name{iniTable}
\alias{iniTable}
\title{Generates initial table to start the simulation (internal function)}
\usage{
iniTable(
  init.individuals,
  init.structure,
  prefix.host,
  ParamHost,
  current.environmental.value = NULL,
  current.cell.number.raster = NULL,
  current.count.A = integer(0),
  current.count.B = integer(0)
)
}
\arguments{
\item{init.individuals}{number of initially infected individuals (i.e. number of lines at time 0).}

\item{init.structure}{State (or coordinates) of the initially infected individuals.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{ParamHost}{list of individual based parameters.}

\item{current.environmental.value}{current value of the environmental  variable provided by the raster according to its position in init.structure.}

\item{current.cell.number.raster}{unique number of the raster cell where the host is}

\item{current.count.A}{current count of host A}

\item{current.count.B}{current count of host B}
}
\description{
This function creates the initial table for the host.
}
\keyword{internal}
