% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetTransmit.R
\name{meetTransmit}
\alias{meetTransmit}
\title{Meet Transmit function}
\usage{
meetTransmit(res, pres.time, positions, timeContactParsed, pTransParsed)
}
\arguments{
\item{res}{an object of class \code{nosoiSim}.}
}
\value{
The modified object res
}
\description{
Perform the tasks related to meeting and transmission events.
}
\keyword{internal}
