% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_step2-MeetTransmit.R
\name{meetTransmit}
\alias{meetTransmit}
\title{Meet Transmit function}
\usage{
meetTransmit(res, pres.time, positions, nContactParsed, pTransParsed)
}
\arguments{
\item{res}{an object of class \code{nosoiSimOne}.}
}
\value{
A table with active transmission events
}
\description{
Perform the tasks related to meeting and transmission events.
}
\keyword{internal}
