% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newLine.R
\name{newLine}
\alias{newLine}
\title{Creates a new line to be added to the table when new host is infected (internal fonction)}
\usage{
newLine(hosts.ID, infected.by, infected.in, time.is, param.pExit,
  param.pMove, param.timeContact, param.pTrans, param.moveDist,
  current.environmental.value = NA)
}
\arguments{
\item{hosts.ID}{unique ID for the new host}

\item{infected.by}{unique ID of host that transmits to the new one}

\item{infected.in}{state in which the host was infected}

\item{time.is}{time in the simulation, when the infection takes place}

\item{param.pExit}{list of exit probability function(s).}

\item{param.pMove}{list of movement probability function(s).}

\item{param.pTrans}{list of transmission probability function(s).}

\item{current.in}{state in which the host currently is}
}
\value{
a list with the new line to add.
}
\description{
This function creates a new line for the table.
The lines are to be bounded with \code{\link[data.table]{rbindlist}}.
}
\keyword{internal}
