% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newLine.R
\name{newLine}
\alias{newLine}
\title{Creates a new line to be added to the table when new host is infected (internal fonction)}
\usage{
newLine(hosts.ID, infected.by, time.is, n.pTrans.param, pTrans.param)
}
\arguments{
\item{hosts.ID}{unique ID for the new host}

\item{infected.by}{unique ID of host that transmits to the new one}

\item{time.is}{time in the simulation, when the infection takes place}

\item{n.pTrans.param}{number of parameters for transmission probability computation (impacts number of columns).}

\item{pTrans.param}{transmission probability function parameters' name.}
}
\description{
Creates a new line to be added to the table when new host is infected (internal fonction)
}
\details{
This function creates the initial table for the host, with 5+number of parameters of the transmission probability function paramters, and init.individuals row(s).
}
