% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoiSim.R
\name{nosoiSim}
\alias{nosoiSim}
\title{Top-level function to use Nosoi}
\usage{
nosoiSim(type = "single", structure = FALSE, ...)
}
\arguments{
\item{type}{specifies which type of pathogen we are interested in, either "single" or "dual"-host (e.g. arboviruses).}

\item{structure}{specifies if the population in which the transmission is to occur is structured (discrete variable).}

\item{...}{arguments to be passed on to the simulator (see below).}
}
\description{
Top-level function to use Nosoi
}
\details{
This function determines which general settings the user wants to use for his simulation.

All other arguments are passed down to the chosen simulator itself, such as \code{\link{singleNone}}
}
