% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{nosoiSimConstructor}
\alias{nosoiSimConstructor}
\title{nosoiSim Constructor}
\usage{
nosoiSimConstructor(N.infected, total.time, table.hosts, table.state,
  prefix.host, type = c("singleNone", "singleDiscrete",
  "singleContinuous", "dualNone", "dualDiscrete", "dualContinuous"))
}
\arguments{
\item{N.infected}{number of infected hosts}

\item{table.hosts}{data.table of hosts}

\item{table.state}{data.table of hosts movement}

\item{type}{name of the diffusion}

\item{pres.time}{current time of the simulation}
}
\value{
An object of class \code{nosoiSim}
}
\description{
Creates a \code{nosoiSim} object.
}
\keyword{internal}
