% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoiSim.R
\name{nosoiSimConstructor}
\alias{nosoiSimConstructor}
\title{nosoiSim Constructor}
\usage{
nosoiSimConstructor(
  total.time,
  type = c("single", "dual"),
  pop.A,
  pop.B = NULL
)
}
\arguments{
\item{type}{population structure (one of "single or "dual)}

\item{pop.A}{an object of class \code{nosoiSimOne} for population A}

\item{pop.B}{an object of class \code{nosoiSimOne} for population B}

\item{pres.time}{current time of the simulation}
}
\value{
An object of class \code{nosoiSim}
}
\description{
Creates a \code{nosoiSim} object.
}
\keyword{internal}
