% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{outputWrapper}
\alias{outputWrapper}
\title{Output wrapper}
\usage{
outputWrapper(Host.count, pres.time, table.hosts, state.archive)
}
\arguments{
\item{Host.count}{number of infected hosts}

\item{pres.time}{current time of the simulation}

\item{table.hosts}{data.table of hosts}

\item{state.archive}{data.table of hosts movement}
}
\description{
Wraps up the nosoi output
}
\keyword{internal}
