% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{paramConstructor}
\alias{paramConstructor}
\title{Param concatenator}
\usage{
paramConstructor(param.pExit, param.pMove, param.timeContact, param.pTrans,
  param.moveDist)
}
\arguments{
\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{param.pMove}{parameter names (list of functions) for the pMove.}

\item{param.timeContact}{parameter names (list of functions) for timeContact.}

\item{param.pTrans}{parameter names (list of functions) for the pTrans.}

\item{param.moveDist}{parameter names (list of functions) for moveDist.}
}
\value{
An object of class \code{ParamHost}
}
\description{
Creates a \code{ParamHost} object.
}
\keyword{internal}
