% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{paramConstructor}
\alias{paramConstructor}
\title{Param concatenator}
\usage{
paramConstructor(param.pExit, param.pMove, param.nContact, param.pTrans,
  param.sdMove)
}
\arguments{
\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{param.pMove}{parameter names (list of functions) for the pMove.}

\item{param.nContact}{parameter names (list of functions) for param.nContact.}

\item{param.pTrans}{parameter names (list of functions) for the pExit.}

\item{param.sdMove}{parameter names (list of functions) for sdMove.}
}
\value{
An object of class \code{ParamHost}
}
\description{
Creates a \code{ParamHost} object.
}
\keyword{internal}
