% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_utilityFunctions.R
\name{parseFunction}
\alias{parseFunction}
\title{Parse function for later use}
\usage{
parseFunction(
  pFunc,
  param.pFunc,
  name,
  diff = FALSE,
  timeDep = FALSE,
  hostCount = FALSE,
  continuous = FALSE,
  stateNames = NA
)
}
\arguments{
\item{pFunc}{a function}

\item{param.pFunc}{a named list of arguments}

\item{name}{the name of the function}

\item{diff}{is the function differential according to state/env.variable? (TRUE/FALSE)}

\item{timeDep}{is the function differential according to absolute time? (TRUE/FALSE)}

\item{hostCount}{is the function differential according to host count? (TRUE/FALSE)}

\item{continuous}{is the function to be used in a continuous space? (TRUE/FALSE)}

\item{stateNames}{name of the states (vector) in case of discrete structure.}
}
\value{
list of parsed quantities:
\itemize{
 \item "vect" Vectorized version of the function.
 \item "vectArgs" Vector of arguments for the vectorized function.
}
}
\description{
Parse a user-provided function to get the vectorized version.
}
\keyword{internal}
