% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks-Messages.R
\name{progressMessage}
\alias{progressMessage}
\title{Progress bar}
\usage{
progressMessage(Host.count.A, Host.count.B = NULL, pres.time, print.step,
  length.sim, max.infected.A, max.infected.B = NULL, type = "single")
}
\arguments{
\item{Host.count.A}{number of infected hosts of host A.}

\item{Host.count.B}{number of infected hosts of host B.}

\item{pres.time}{current time of the simulation}

\item{print.step}{if print.progress is TRUE, step with which the progress message will be printed.}

\item{length.sim}{the length (in unit of time) over which the simulation should be run.}

\item{max.infected.A}{the maximum number of hosts that can be infected in the simulation for host A.}

\item{max.infected.B}{the maximum number of hosts that can be infected in the simulation for host B.}

\item{type}{either single/dual host}
}
\description{
Echos the state of the simulation at any given time step provided by the user.
}
\keyword{internal}
