% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosoi_sanityChecks-Messages.R
\name{RasterSanityChecks}
\alias{RasterSanityChecks}
\title{Checks if the raster is properly formatted}
\usage{
RasterSanityChecks(structure.raster, init.structure, none.at.start = NULL)
}
\arguments{
\item{structure.raster}{raster object defining the environmental variable.}

\item{init.structure}{which state (i.e. location) the initially infected individuals are located.}
}
\description{
Checks if the environmental raster is properly formatted given the user's input and nosoi's requirements.
}
\keyword{internal}
