% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{sampleTransmissionTree}
\alias{sampleTransmissionTree}
\title{Sample the transmission tree}
\usage{
sampleTransmissionTree(nosoiInf, tree, samples)
}
\arguments{
\item{nosoiInf}{an object of class \code{\link{nosoiSim}}}

\item{tree}{a \code{\link[tidytree:treedata]{treedata}} object created by function \code{\link{getTransmissionTree}}}

\item{samples}{a \code{\link[data.table:data.table-package]{data.table}} object with the following entries:
\describe{
  \item{hosts}{Host ID of the individuals to be sampled}
  \item{times}{Times at which each host is sampled}
  \item{labels}{label for the corresponding tip in the tree}
}}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree and the mapped data at all the nodes.
}
\description{
Sample a full transmission tree. This function allows for sampling multiple
 times on the same lineage. When this happens, the sampled anscestor is
 a tip with length zero.
}
\details{
The tree needs to be produced by function \code{\link{getTransmissionTree}}
 applied on the same \code{nosoiSim} object.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

For sampling only dead individuals, see \code{\link{sampleTransmissionTreeFromExiting}}.
}
