% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{sampleTransmissionTreeFromExiting}
\alias{sampleTransmissionTreeFromExiting}
\title{Sample the transmission tree among the exited hosts}
\usage{
sampleTransmissionTreeFromExiting(tree, hosts)
}
\arguments{
\item{tree}{a \code{\link[tidytree:treedata-class]{treedata}} object created by function \code{\link{getTransmissionTree}}}

\item{hosts}{a vector of dead hosts to sample}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree and the mapped data at all the nodes.
}
\description{
Sample a full transmission tree. This function allows for sampling only dead
 individuals (e.g. when the sampling procedure is destructive).
 Beware because it does not influence the epidemiological process, it only means that the host
 has been sampled when exiting the simulation
}
\details{
The tree needs to be produced by function \code{\link{getTransmissionTree}}
 applied on the same \code{nosoiSim} object.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

For sampling non-dead individuals, see \code{\link{sampleTransmissionTree}}.
}
