% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_treeGenerator.r
\name{sampleTransmissionTreeFromExiting}
\alias{sampleTransmissionTreeFromExiting}
\title{Samples the transmission tree (phylogenetic tree-like) among the exited hosts}
\usage{
sampleTransmissionTreeFromExiting(tree, hosts)
}
\arguments{
\item{tree}{a \code{\link[tidytree:treedata-class]{treedata}} object created by function \code{\link{getTransmissionTree}}}

\item{hosts}{a vector of dead hosts to sample}
}
\value{
A tree of class \code{\link[tidytree:treedata-class]{treedata}}, containing a
phylogenetic tree based on the transmission chain and the mapped data at all the nodes.
}
\description{
Sample a full transmission tree. This function allows for sampling only exited (i.e. inactive)
 individuals (e.g. when the sampling procedure is destructive or cuts the hosts from the population).
 Beware because it does not influence the epidemiological process, it only means that the host
 has been sampled when exiting the simulation.
}
\details{
The tree needs to be produced by function \code{\link{getTransmissionTree}}
 applied on the same \code{nosoiSim} object.
}
\seealso{
For exporting the annotated tree to other softwares, see functions
in \pkg{treeio} (e.g. \code{\link[treeio:write.beast]{write.beast}}).

For sampling non-dead individuals, see \code{\link{sampleTransmissionTree}}.
}
