% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-continuous.R
\name{singleContinuous}
\alias{singleContinuous}
\title{Single-host with structured host population}
\usage{
singleContinuous(type, structure, length.sim, max.infected,
  init.individuals, init.structure, structure.raster, diff.pMove = FALSE,
  timeDep.pMove = FALSE, pMove, param.pMove, diff.moveDist = FALSE,
  timeDep.moveDist = FALSE, moveDist, param.moveDist,
  attracted.by.raster = FALSE, diff.timeContact = FALSE,
  timeDep.timeContact = FALSE, timeContact, param.timeContact,
  diff.pTrans = FALSE, timeDep.pTrans = FALSE, pTrans, param.pTrans,
  diff.pExit = FALSE, timeDep.pExit = FALSE, pExit, param.pExit,
  prefix.host = "H", progress.bar = TRUE, print.step = 10, ...)
}
\arguments{
\item{type}{single host or dual host}

\item{structure}{is there any discrete structure in the host population to be accounted for (such as spatial features)}

\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{number of initially infected individuals.}

\item{init.structure}{which state (i.e. location) the initially infected individuals are located.}

\item{structure.raster}{raster object defining the environmental variable.}

\item{diff.pMove}{is pMove different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pMove}{is pMove dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pMove}{function that gives the probability of a host moving as a function of time.}

\item{param.pMove}{parameter names (list of functions) for the pMove.}

\item{diff.moveDist}{is moveDist dependant on the environmental value (TRUE/FALSE).}

\item{timeDep.moveDist}{is moveDist dependant on the absolute time of the simulation (TRUE/FALSE).}

\item{moveDist}{function that gives the distance travelled (based on coordinates); gives the sd value for the brownian motion.}

\item{param.moveDist}{parameter names (list of functions) for moveDist.}

\item{attracted.by.raster}{should the hosts be attracted by high values in the environmental raster? (TRUE/FALSE)}

\item{diff.timeContact}{is timeContact different between states of the structured population (TRUE/FALSE)}

\item{timeDep.timeContact}{is timeContact dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{timeContact}{function that gives the number of potential transmission events per unit of time.}

\item{param.timeContact}{parameter names (list of functions) for timeContact.}

\item{diff.pTrans}{is pTrans different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pTrans}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pTrans}{function that gives the probability of transmit a pathogen as a function of time since infection.}

\item{param.pTrans}{parameter names (list of functions) for the pTrans.}

\item{diff.pExit}{is pExit different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pExit}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{pExit}{function that gives the probability to exit the simulation for an infected host (either moving out, dying, etc.).}

\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{progress.bar}{if TRUE, print message on simulation state to screen.}

\item{print.step}{progress.bar is TRUE, step with which the progress message will be printed.}

\item{...}{other arguments to be passed on to the simulator (see below).}
}
\description{
This function runs a single-host transmission chain simulation, with a structured host population (such as spatial features).
The simulation stops either at the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected)
is passed.
}
