% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-discrete.R
\name{singleDiscrete}
\alias{singleDiscrete}
\title{Single-host with structured (discrete) host population}
\usage{
singleDiscrete(length.sim, max.infected, init.individuals, init.structure,
  structure.matrix, diff.pMove = FALSE, timeDep.pMove = FALSE,
  hostCount.pMove = FALSE, pMove, param.pMove, diff.nContact = FALSE,
  timeDep.nContact = FALSE, hostCount.nContact = FALSE, nContact,
  param.nContact, diff.pTrans = FALSE, timeDep.pTrans = FALSE,
  hostCount.pTrans = FALSE, pTrans, param.pTrans, diff.pExit = FALSE,
  timeDep.pExit = FALSE, hostCount.pExit = FALSE, pExit, param.pExit,
  prefix.host = "H", print.progress = TRUE, print.step = 10)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{number of initially infected individuals.}

\item{init.structure}{which state (e.g. location) the initially infected individuals are located.}

\item{structure.matrix}{transition matrix (probabilities) to go from location A (row) to B (column)}

\item{diff.pMove}{is pMove different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pMove}{is pMove dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{hostCount.pMove}{does pMove varies with the host count in the state? (TRUE/FALSE); diff.pMove should be TRUE.}

\item{pMove}{function that gives the probability of a host moving as a function of time.}

\item{param.pMove}{parameter names (list of functions) for the pMove.}

\item{diff.nContact}{is nContact different between states of the structured population (TRUE/FALSE)}

\item{timeDep.nContact}{is nContact dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{hostCount.nContact}{does nContact varies with the host count in the state? (TRUE/FALSE); diff.nContact should be TRUE.}

\item{nContact}{function that gives the number of potential transmission events per unit of time.}

\item{param.nContact}{parameter names (list of functions) for nContact.}

\item{diff.pTrans}{is pTrans different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pTrans}{is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{hostCount.pTrans}{does pTrans varies with the host count in the state? (TRUE/FALSE); diff.pTrans should be TRUE.}

\item{pTrans}{function that gives the probability of transmit a pathogen as a function of time since infection.}

\item{param.pTrans}{parameter names (list of functions) for the pTrans.}

\item{diff.pExit}{is pExit different between states of the structured population (TRUE/FALSE)}

\item{timeDep.pExit}{is pExit dependant on the absolute time of the simulation (TRUE/FALSE)}

\item{hostCount.pExit}{does pExit varies with the host count in the state? (TRUE/FALSE); diff.pExit should be TRUE.}

\item{pExit}{function that gives the probability to exit the simulation for an infected host (either moving out, dying, etc.).}

\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{print.progress}{if TRUE, displays a progress bar (current time/length.sim).}

\item{print.step}{print.progress is TRUE, step with which the progress message will be printed.}
}
\description{
This function runs a single-host transmission chain simulation, with a structured host population (such as spatial features).
The simulation stops either at the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected)
is passed.
}
