% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-none.R
\name{singleNone}
\alias{singleNone}
\title{Single-host without structured host population}
\usage{
singleNone(length.sim, max.infected, init.individuals, timeContact,
  param.timeContact, pTrans, param.pTrans, pExit, param.pExit,
  prefix.host = "H", progress.bar = TRUE, ...)
}
\arguments{
\item{length.sim}{specifies the length (in unit of time) over which the simulation should be run.}

\item{max.infected}{specifies the maximum number of hosts that can be infected in the simulation.}

\item{init.individuals}{number of initially infected individuals.}

\item{timeContact}{function that gives the number of potential transmission events per unit of time.}

\item{param.timeContact}{parameter names (list of functions) for param.timeContact.}

\item{pTrans}{function that gives the probability of transmit a pathogen as a function of time since infection.}

\item{param.pTrans}{parameter names (list of functions) for the pExit.}

\item{pExit}{function that gives the probability to exit the simulation for an infected host (either moving out, dying, etc.).}

\item{param.pExit}{parameter names (list of functions) for the pExit.}

\item{prefix.host}{character(s) to be used as a prefix for the hosts identification number.}

\item{progress.bar}{if TRUE, displays a progress bar (current time/length.sim).}

\item{...}{other arguments to be passed on to the simulator (see below).}
}
\description{
This function runs a single-host transmission chain simulation, without any spatial features. The simulation stops either at
the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.
}
